requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
Flask-Session==0.5.0
fonttools==4.40.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql-connector-python==8.0.33
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.3.4
